requirements.txt
flask
torch
torchvision
numpy
Pillow
scikit-learn
pandas
tqdm
open_clip_torch

